MyStudyApp/requirements.txt
asgiref==3.8.1
click==8.2.1
dj-database-url==3.0.1
Django==5.2.2
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
uvicorn==0.35.0
whitenoise==6.9.0
